better default mac framework ruby install locations
Trying to mimic the behavior of the mac framework installed Rubygems
as closely as possible here. /Library/Ruby/RPG/1.8 is used to store
database information, executables are installed to /usr/bin, and
library files are installed to /usr/lib/ruby/vendor_ruby/1.8.

On Snow Leopard, this ALMOST doesn't require root privs. The
/Library/Ruby and /usr/lib/ruby hierarchies are group writeable and
group owned by "admin". The /usr/bin path is the only location that
requires root/wheel. Rubygems falls back to installing to ~/.gem
when all locations are not writeable. We may need to consider
adding something similar.