Add --rpgdir option to configure for rpg database and installed packages.
Previously rpg would always try to install packages into a ruby
libdir subdirectory. This required rpg to run with root privileges
if there was only one ruby (the system one) installed.

By specifying --rpgdir it is possible to locate rpg database and
installed packages anywhere on the filesystem, allowing rpg to be
run as an unprivileged user on systems with a single ruby.